Updates and fix saving of histograms.
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@192257 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 056e8f8fd5ca380f4cbc6b0a2092186cab05c6a6

+++ Online/Online/Hlt2Monitoring/Hlt2Monitoring/Common.h
@@ -5,7 +5,7 @@ namespace Monitoring {
 
 using TCK = unsigned int;
 using HistId = long unsigned int;
-using RunNumber = unsigned long;
+using RunNumber = unsigned int;
 using BinContent = unsigned long long;
 using Seconds = std::size_t;
